meta-mender-chip: add integration for NTC C.H.I.P.
The C.H.I.P. and C.H.I.P. Pro are the only two boards that NextThingCo
was able to produce before their tragic demise. They are neat little
boards that offer wifi connected tomfoolery.

Unfortunately, not much support for them reached mainline u-boot or
Linux. U-boot is stuck at 2016.1 and Linux at 4.4.30. The Linux kernel
works more or less as is, but the u-boot require some patching to work.

This integration is based on my fork of joelguittet/meta-chip:

- https://github.com/Risca/meta-chip

It focuses on the C.H.I.P., but with some more elbow grease it should
work fine for the C.H.I.P. Pro as well.

The meta-chip layer does the minimum necessary to make everything build
again, and the meta-mender-chip layer in this commit further patches
u-boot so it can support Mender. The current Mender u-boot patches apply
with some fuzz, but they seem to work. The kernel didn't need patching
at this point.

The C.H.I.P. come with a rather odd MLC NAND chip where none of the
default NAND parameters were right and had to be explicitly set. The
flash partitioning is as follows:

    0x000000000000-0x000000400000 : "spl"
    0x000000400000-0x000000800000 : "spl-backup"
    0x000000800000-0x000000c00000 : "u-boot"
    0x000000c00000-0x000001400000 : "u-boot-env"
    0x000001400000-0x000100000000 : "ubi"

Where "spl" is a first stage bootloader and its backup. The rest should
be fairly self explanatory: "u-boot" for u-boot, "u-boot-env" for its
environment, and "ubi" for rootfsa, rootfsb, and data volumes. The
u-boot environment at build time is also present as two volumes. I
couldn't find a way to patch that away, and I should probably not try to
either. Future editions of this port may move the bootloader environment
to the ubi volumes instead.

Signed-off-by: Patrik Dahlström <risca@dalakolonin.se>